feat(config env): Add config environment defaults
This commit refactors the witnet_config codebase in order to support
different default constants and parameters depending on the
environment in which the node is going to run (tesnet, mainnet, etc).

It does so by separating the the configuration used in the
application:

`config::Config`

from the one that is used as temporary container of config parameters
loaded from configuration files or command line arguments:

`config::partial::Config`

All the loaders defined in `loaders` will return a
`config::partial::Config` object, then that object will be consumed
by:

`config::Config::from_partial`

in order to merge it with the declared environment
defaults (`defaults` module).

Example:

As a complete example suppose you want to build a config object
containing all the defaults for mainnet, you'll do it like this:

```
use witnet_config::config::{partial, Config, Environment};

fn get_mainnet_config() -> Config {
    Config::from_partial(partial::Config::default_mainnet())
}
```

Improvements:

Currently there's a lot of duplication between `Config` and
`partial::Config`, the later is just the same as the former but with
all the fields marked as optional `Option`, this is a good candidate
for being generated by a macro. I found a crate that could help as a
starting point:

https://github.com/pLesur/OptionalStruct

the reason why I did not used it was because it does not play nice
with other derive attributes (issue #8) and because is gives no chance
of hiding the auto-generated `apply_options` method.