fix(node): send PeersBeacons even if peers < outbound limit
This fixes a bug where the node would move to WaitingConsensus as soon
as one of its peers closed the connection.

Now the SessionsManager will always send the list of beacons to the
ChainManager. If there is consensus after counting the missing peers as
"NO BEACON", the node will start synchronizing even if it's still
missing some peers.